% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amount_donated_best_stan.R
\name{amount_donated_best}
\alias{amount_donated_best}
\title{amount_donated_best}
\usage{
amount_donated_best(df, ...)
}
\arguments{
\item{df}{Data frame with two columns: amount (numeric) and group (factor)}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Stan implementation of John Kruschke's Bayesian Estimation Supersedes the
t-test (BEST).
}
