% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiled_models.R
\name{compiled_models}
\alias{compiled_models}
\alias{donate_likely}
\alias{amount_donated_best}
\title{Stan models}
\usage{
donate_likely()

amount_donated_best()
}
\description{
Precompiled Stan models ready to be used with \code{rstan::sampling}
}
\details{
Instead of writing wrapper functions for these two models, we just provide them as precompiled models that we can then use on their own
}
