% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crackdownsphilanthropy-package.R
\docType{package}
\name{crackdownsphilanthropy-package}
\alias{crackdownsphilanthropy-package}
\alias{crackdownsphilanthropy}
\title{The 'crackdownsphilanthropy' package.}
\description{
Research on how anti-NGO crackdowns influence donor behavior
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
all
}
