% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donate_likely_stan.R
\name{donate_likely}
\alias{donate_likely}
\title{donate_likely_stan}
\usage{
donate_likely(n_total_1, n_total_2, n_likely_1, n_likely_2, ...)
}
\arguments{
\item{n_total_1}{Total number in group 1}

\item{n_total_2}{Total number in group 2}

\item{n_likely_1}{Number of answers responding "Likely" in group 1}

\item{n_likely_2}{Number of answers responding "Likely" in group 1}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Model the difference between two proportions. Essentially a Bayesian version
of Fisher's exact test.
}
