% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_quantiles.R
\name{factor_quantiles}
\alias{factor_quantiles}
\alias{quantile.ordered}
\alias{quantile.factor}
\alias{median.ordered}
\alias{median.factor}
\title{Factor quantiles}
\source{
Borrowed from Jeffrey Arnold: \url{https://gist.github.com/jrnold/6759254}
}
\usage{
\method{quantile}{ordered}(x, probs = seq(0, 1, 0.25))

\method{quantile}{factor}(x, probs = seq(0, 1, 0.25))

\method{median}{ordered}(x)

\method{median}{factor}(x)
}
\arguments{
\item{x}{factor to calculate quantiles for}

\item{probs}{numeric vector of probabilities}
}
\description{
Calculate the quantiles of factor variables.
}
\examples{
# Normal factors
responses <- factor(c("Agree", "Agree", "Agree", "Disagree", "Disagree"))
median(responses)
quantile(responses, probs = c(0.25, 0.75))

# Ordered factors
responses_ordered <- factor(c("Agree", "Agree", "Agree", "Disagree", "Disagree"),
                            ordered = TRUE)
median(responses_ordered)
quantile(responses_ordered, probs = c(0.25, 0.75))

# In the case of ties, the first in order will win.
with_ties <- factor(c("A", "A", "B", "B"), ordered = TRUE)
median(with_ties)  # "A"
}
