% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{fmt_seconds}
\alias{fmt_seconds}
\title{Format seconds nicely}
\usage{
fmt_seconds(x)
}
\arguments{
\item{x}{a numeric value (in seconds)}
}
\value{
A string value (in HH:MM format)
}
\description{
Convert a given number of seconds into "HH:MM" format
}
\examples{
fmt_seconds(5)
#> "00:05"

fmt_seconds(500)
#> "08:20"
}
