% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{data}
\name{ngo_colors}
\alias{ngo_colors}
\title{Hex codes for colors used in project}
\format{An object of class \code{character} of length 7.}
\usage{
ngo_colors
}
\description{
Hex codes for colors used in project
}
\keyword{datasets}
