% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ngo_cols}
\alias{ngo_cols}
\title{Function to extract NGO colors as hex codes}
\usage{
ngo_cols(..., name = TRUE)
}
\arguments{
\item{...}{Character names of ngo_colors}
}
\description{
Function to extract NGO colors as hex codes
}
