% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{data}
\name{ngo_palettes}
\alias{ngo_palettes}
\title{Palette definitions}
\format{An object of class \code{list} of length 5.}
\usage{
ngo_palettes
}
\description{
Palette definitions
}
\keyword{datasets}
