% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_prop_stan.R
\name{pop_prop_stan}
\alias{pop_prop_stan}
\title{pop_prop_stan}
\usage{
pop_prop_stan(x, n_total, pop_prop, ...)
}
\arguments{
\item{x}{Number of observations in sample with characteristic}

\item{n_total}{Number of observations in sample}

\item{pop_prop}{Population proportion to compare with (x / n_total)}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Compare the proportion of a characteristic in a sample with a population-level proportion.
}
