% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\alias{md_summary_row}
\title{Generate a summary table with sparkgraphs}
\usage{
md_summary_row(value, variable_name)
}
\arguments{
\item{value}{a numeric or factor vector}

\item{variable_name}{name of the variable to summarize (generally the column name of value)}
}
\description{
Create a fancy summary Markdown table for numeric and factor variables with tiny ggplot sparkgraphs
}
\examples{
\dontrun{
set.seed(1234)
example_data <- tibble(numeric_thing = 1:100,
                       factor_thing = factor(sample(LETTERS[1:4], 100, 
                                                    replace = TRUE)))

# Single rows
md_summary_row(example_data[[1]], "numeric_thing")
#> # A tibble: 1 x 2
#>   spark                                           summary
#>   <chr>                                           <chr>
#> 1 ![](output/figures/summary-table/numeric_thing) Median: 50.5 | Mean: 50.5…

# Make a summary for all rows
example_data \%>\% 
  summarise_all(~list(.)) \%>\% # Collapse each column into a nested cell
  pivot_longer(everything(), names_to = "variable_name", values_to = "value") \%>\% 
  mutate(summary = map2(.x = value, .y = variable_name, 
                        ~ md_summary_row(.x, .y))) \%>\% 
  select(-value) \%>\% 
  unnest(summary)
#> # A tibble: 2 x 3
#>   variable      spark                                   summary
#>   <chr>         <chr>                                   <chr>
#> 1 numeric_thing ![](output/figures/summary-table/numer… Median: 50.5 | Mean…
#> 2 factor_thing  ![](output/figures/summary-table/facto… A (19; 19.0\%) | B (…
}
}
