% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{theme_ngos}
\alias{theme_ngos}
\title{theme_ngos}
\usage{
theme_ngos(base_size = 11, base_family = "Roboto Condensed", density = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family (default is Roboto Condensed)}

\item{density}{Binary indicator to remove the panel grid and make a few other adjustments}
}
\description{
A custom ggplot2 theme used throughout this project
}
\examples{
library(ggplot2)

# Standard theme
ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_ngos()

# using density = TRUE
ggplot(mtcars, aes(x = mpg)) +
  geom_density(fill = "grey80") +
  theme_ngos(density = TRUE)
}
