% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{data}
\name{theme_ngos_table}
\alias{theme_ngos_table}
\title{theme_ngos_table}
\format{
An object of class \code{list} of length 3.
}
\usage{
theme_ngos_table
}
\description{
A custom ggplot2 theme used for \code{tableGrob} objects`
}
\examples{
library(grid)
d <- head(iris, 3)
g <- tableGrob(d, rows = NULL, theme = theme_ngos_table)
grid.newpage()
grid.draw(g)
}
\keyword{datasets}
